chore(main): move packages use repo to common
