+++ src/main/java/com/example/cheerdo/entity/Todo.java
@@ -2,7 +2,6 @@ package com.example.cheerdo.entity;
 
 import com.example.cheerdo.todo.dto.response.TodoResponseDto;
 import com.example.cheerdo.todo.enums.Type;
-import com.example.cheerdo.todo.repository.TodoRepository;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
