+++ src/main/java/com/example/cheerdo/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.example.cheerdo.post.repository;
+package com.example.cheerdo.repository;
 
 import com.example.cheerdo.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
