+++ src/main/java/com/example/cheerdo/repository/RelationRepository.java
@@ -1,4 +1,4 @@
-package com.example.cheerdo.post.repository;
+package com.example.cheerdo.repository;
 
 import com.example.cheerdo.entity.FriendRelation;
 import com.example.cheerdo.entity.Member;
