+++ src/main/java/com/example/cheerdo/todo/service/TodoServiceImpl.java
@@ -8,14 +8,13 @@ import com.example.cheerdo.todo.dto.request.WriteTodoRequestDto;
 import com.example.cheerdo.todo.dto.response.TodoResponseDto;
 import com.example.cheerdo.todo.enums.Type;
 import com.example.cheerdo.todo.repository.MemberRepository;
-import com.example.cheerdo.todo.repository.TodoRepository;
+import com.example.cheerdo.repository.TodoRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.text.MessageFormat;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
