+++ src/test/java/com/example/cheerdo/post/service/PostServiceImplTest.java
@@ -5,9 +5,9 @@ import com.example.cheerdo.entity.Member;
 import com.example.cheerdo.post.dto.request.LetterRequestDto;
 import com.example.cheerdo.post.dto.request.PostRequestDto;
 import com.example.cheerdo.post.dto.response.PostResponseDto;
-import com.example.cheerdo.post.repository.MemberRepository;
-import com.example.cheerdo.post.repository.PostRepository;
-import com.example.cheerdo.post.repository.RelationRepository;
+import com.example.cheerdo.repository.MemberRepository;
+import com.example.cheerdo.repository.PostRepository;
+import com.example.cheerdo.repository.RelationRepository;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
